Move IAnimatedBlockFactory into appropriate package
