+++ bigdoors-core/src/main/java/nl/pim16aap2/bigdoors/managers/AnimatedBlockHookManager.java
@@ -1,10 +1,10 @@
 package nl.pim16aap2.bigdoors.managers;
 
 import nl.pim16aap2.bigdoors.api.animatedblockhook.IAnimatedBlockHook;
-import nl.pim16aap2.bigdoors.api.animatedblockhook.IAnimatedBlockHookFactory;
 import nl.pim16aap2.bigdoors.api.animatedblock.IAnimatedBlock;
 import nl.pim16aap2.bigdoors.api.debugging.DebuggableRegistry;
 import nl.pim16aap2.bigdoors.api.debugging.IDebuggable;
+import nl.pim16aap2.bigdoors.api.factories.IAnimatedBlockHookFactory;
 import nl.pim16aap2.util.SafeStringBuilder;
 
 import javax.inject.Inject;
